requirements.txt
celery
python-dotenv